[fix] update dhcpd3 munin plugin configuration.
In Debian Squeeze, the ISC DHCP server package is now named
isc-dhcp-server and the configuration and lease files have slightly
different path.

The dhcpd3 munin plugin should still be usable even though the dhcp
server version is now > 4.

That said, since the dhcpd3 munin plugin is in the munin-plugins-extra
and this package conflicts with pf-xivo-monitoring, there's still some
work to do before it's working.