don't check if asterisk is deadlocked/frozen
The res_freeze_check module now does it all by itself, i.e. it abort the
process if it detects that it is deadlocked.

So "revert" the check to the state it was before res_freeze_check, i.e.
just check if we can login to the AMI or not.