monit: only check asterisk with pidfile
Remove the AMI check, mostly because it's a bit fragile: even when
asterisk is fine, the check can fail if monit does it's "expect" on the
AMI banner instead of the AMI response. It's also a bit unstable:
recently, on some installation, sometimes connection are not accepted
successfully via the AMI for some unknown reason (we are currently trying
to understand why).

The check doesn't add that much value over just checking if the process
is alive with the help of the pidfile.